Address part of reviewers' comments
