+++ junit4/build.gradle
@@ -8,4 +8,11 @@ task generateSources(type: Copy) {
     into "${project.ext.genSrcDir}/main/java/com/linecorp/armeria/internal/testing"
 }
 
+// Copy shared files from junit5 module to junit4 module.
+task copyJunitSources(type: Copy) {
+    from "${rootProject.projectDir}/junit5/src/main/java/com/linecorp/armeria/testing/server/ServiceRequestContextCaptor.java"
+    into "${project.ext.genSrcDir}/main/java/com/linecorp/armeria/testing/server"
+}
+
 tasks.compileJava.dependsOn(generateSources)
+tasks.compileJava.dependsOn(copyJunitSources)
